Much nicer debug output

+++ system/modules/backend/Automator.php
@@ -194,7 +194,7 @@ class Automator extends Backend
 		{
 			foreach ($arrScripts as $strFile)
 			{
-				if ($strFile != 'index.html' && $strFile != 'contao.css' && !is_dir(TL_ROOT . '/system/scripts/' . $strFile))
+				if ($strFile != 'index.html' && $strFile != 'contao.css' && $strFile != 'debug.css' && !is_dir(TL_ROOT . '/system/scripts/' . $strFile))
 				{
 					unlink(TL_ROOT . '/assets/css/' . $strFile);
 				}
